refactor: 스크리닝 리뷰 리스트 페이지 생성
